+++ src/com/lukxtech/powertrain/clutchfamily/clutches/hydraulic/exceptions/InvalidHydraulicFluidTypeException.java
@@ -1,0 +1,7 @@
+package com.lukxtech.powertrain.clutchfamily.clutches.hydraulic.exceptions;
+
+public class InvalidHydraulicFluidTypeException extends RuntimeException {
+    public InvalidHydraulicFluidTypeException(String message) {
+        super(message);
+    }
+}