+++ src/com/lukxtech/powertrain/clutchfamily/clutches/platebased/common/exceptions/InvalidNumberOfPlateValueException.java
@@ -1,0 +1,7 @@
+package com.lukxtech.powertrain.clutchfamily.clutches.platebased.common.exceptions;
+
+public class InvalidNumberOfPlateValueException extends RuntimeException {
+    public InvalidNumberOfPlateValueException(String message) {
+        super(message);
+    }
+}