+++ src/com/lukxtech/powertrain/clutchfamily/common/exceptions/BadClutchTypeException.java
@@ -1,0 +1,7 @@
+package com.lukxtech.powertrain.clutchfamily.common.exceptions;
+
+public class BadClutchTypeException extends RuntimeException {
+    public BadClutchTypeException(String message) {
+        super(message);
+    }
+}