+++ src/com/lukxtech/powertrain/clutchfamily/common/exceptions/InvalidActuatorTypeException.java
@@ -1,0 +1,7 @@
+package com.lukxtech.powertrain.clutchfamily.common.exceptions;
+
+public class InvalidActuatorTypeException extends RuntimeException {
+    public InvalidActuatorTypeException(String message) {
+        super(message);
+    }
+}