+++ src/com/lukxtech/powertrain/clutchfamily/common/exceptions/InvalidCoolingTypeException.java
@@ -1,0 +1,7 @@
+package com.lukxtech.powertrain.clutchfamily.common.exceptions;
+
+public class InvalidCoolingTypeException extends RuntimeException {
+    public InvalidCoolingTypeException(String message) {
+        super(message);
+    }
+}