+++ src/com/lukxtech/powertrain/clutchfamily/interfaces/DualClutchInterface.java
@@ -1,0 +1,7 @@
+package com.lukxtech.powertrain.clutchfamily.interfaces;
+
+public interface DualClutchInterface {
+    int getNumberOfClutches();
+    double getEngagementTime();
+    void setEngagementTime();
+}