+++ src/com/lukxtech/powertrain/clutchfamily/interfaces/ElectronicControlBasedClutchInterface.java
@@ -1,0 +1,7 @@
+package com.lukxtech.powertrain.clutchfamily.interfaces;
+
+public interface ElectronicControlBasedClutchInterface {
+    String getControlModuleVersion();
+    void setControlModuleVersion(String ver);
+//    ArrayList<Sensors> getSensorsAvailable;
+}