+++ src/com/lukxtech/powertrain/clutchfamily/interfaces/HydraulicClutchInterface.java
@@ -1,0 +1,9 @@
+package com.lukxtech.powertrain.clutchfamily.interfaces;
+
+import com.lukxtech.powertrain.clutchfamily.clutches.hydraulic.enums.HydraulicFluidType;
+
+public interface HydraulicClutchInterface {
+    HydraulicFluidType getFluidType();
+    void setFluidType(String fluidType);
+    void setFluidType(HydraulicFluidType fluidType);
+}