+++ src/com/lukxtech/powertrain/clutchfamily/interfaces/PlateBasedClutchInterface.java
@@ -1,0 +1,6 @@
+package com.lukxtech.powertrain.clutchfamily.interfaces;
+
+public interface PlateBasedClutchInterface extends ClutchInterface {
+    int getNumberOfPlates();
+    void setNumberOfPlates(int numberOfPlates);
+}